backend/requirement.txt
asgiref==3.7.2
boto3==1.34.11
botocore==1.34.11
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==5.0.1
django-cors-headers==4.3.1
django-ses==3.5.2
django-storages==1.14.2
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
gunicorn==21.2.0
idna==3.6
jmespath==1.0.1
oauthlib==3.2.2
packaging==23.2
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.10.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.0.7
